Amount of stars for starboard now dynamic
Starboard now requires the amount of members in the server divided by 20 rounded to the nearest integer if that amount is greater than 2. Will always require a minimum of 2.
For example, a server with 40 members only needs 2 stars, but a server with 60 needs 3. 71 requires 4.
Formula can be seen on line 72.